Create autogen_to_sql
Use Autogen to write a chat to database agent, and visualize and summarize feature data. For example, if you ask a question like "How many cars in Baiyun District, Guangzhou ran a red light?", it will be translated into relevant SQL to query the table, output the corresponding data, and finally return it to the user.